Refactor NodeNameManager to backends.metadatastore
